n x n Matrix Multiplication Python Code
Solution1: triple nested loop, which time complexity is O(n^3)
Solution2: based on divide and conquer technique, which time complexity is O(n^3).

However, this code demonstrates that even their time complexities are same theoretically, solution1 actual running time is faster than solution2. This is due to the heavy time cost spent when calling the functions and stacking up in the system.